+++ src/request.js
@@ -8,7 +8,7 @@
  */
 
 import {format as formatUrl} from 'url';
-import Headers, {exportNodeCompatibleHeaders} from './headers.js';
+import Headers from './headers.js';
 import Body, {clone, extractContentType, getTotalBytes} from './body.js';
 import {isAbortSignal} from './utils/is.js';
 import {getSearch} from './utils/get-search.js';
@@ -21,12 +21,12 @@ const INTERNALS = Symbol('Request internals');
  * @param  {*} obj
  * @return {boolean}
  */
-function isRequest(object) {
+const isRequest = object => {
 	return (
 		typeof object === 'object' &&
 		typeof object[INTERNALS] === 'object'
 	);
-}
+};
 
 /**
  * Wrapper around `new URL` to handle relative URLs (https://github.com/nodejs/node/issues/12682)
@@ -34,7 +34,7 @@ function isRequest(object) {
  * @param  {string} urlStr
  * @return {void}
  */
-function parseURL(urlString) {
+const parseURL = urlString => {
 	/*
 		Check whether the URL is absolute or not
 
@@ -46,7 +46,7 @@ function parseURL(urlString) {
 	}
 
 	throw new TypeError('Only absolute URLs are supported');
-}
+};
 
 /**
  * Request class
@@ -60,7 +60,9 @@ export default class Request extends Body {
 		let parsedURL;
 
 		// Normalize input and force URL to be encoded as UTF-8 (https://github.com/bitinn/node-fetch/issues/245)
-		if (!isRequest(input)) {
+		if (isRequest(input)) {
+			parsedURL = parseURL(input.url);
+		} else {
 			if (input && input.href) {
 				// In order to support Node.js' Url objects; though WHATWG's URL objects
 				// will fall into this branch also (since their `toString()` will return
@@ -72,28 +74,24 @@ export default class Request extends Body {
 			}
 
 			input = {};
-		} else {
-			parsedURL = parseURL(input.url);
 		}
 
 		let method = init.method || input.method || 'GET';
 		method = method.toUpperCase();
 
 		// eslint-disable-next-line no-eq-null, eqeqeq
-		if ((init.body != null || isRequest(input) && input.body !== null) &&
+		if (((init.body != null || isRequest(input)) && input.body !== null) &&
 			(method === 'GET' || method === 'HEAD')) {
 			throw new TypeError('Request with GET/HEAD method cannot have body');
 		}
 
-		// eslint-disable-next-line no-eq-null, eqeqeq
-		const inputBody = init.body != null ?
+		const inputBody = init.body ?
 			init.body :
 			(isRequest(input) && input.body !== null ?
 				clone(input) :
 				null);
 
 		super(inputBody, {
-			timeout: init.timeout || input.timeout || 0,
 			size: init.size || input.size || 0
 		});
 
@@ -126,12 +124,8 @@ export default class Request extends Body {
 		};
 
 		// Node-fetch-only options
-		this.follow = init.follow !== undefined ?
-			init.follow : (input.follow !== undefined ?
-				input.follow : 20);
-		this.compress = init.compress !== undefined ?
-			init.compress : (input.compress !== undefined ?
-				input.compress : true);
+		this.follow = init.follow === undefined ? (input.follow === undefined ? 20 : input.follow) : init.follow;
+		this.compress = init.compress === undefined ? (input.compress === undefined ? true : input.compress) : init.compress;
 		this.counter = init.counter || input.counter || 0;
 		this.agent = init.agent || input.agent;
 		this.highWaterMark = init.highWaterMark || input.highWaterMark || 16384;
@@ -186,7 +180,7 @@ Object.defineProperties(Request.prototype, {
  * @param   Request  A Request instance
  * @return  Object   The options object to be passed to http.request
  */
-export function getNodeRequestOptions(request) {
+export const getNodeRequestOptions = request => {
 	const {parsedURL} = request[INTERNALS];
 	const headers = new Headers(request[INTERNALS].headers);
 
@@ -218,7 +212,7 @@ export function getNodeRequestOptions(request) {
 
 	// HTTP-network-or-cache fetch step 2.11
 	if (!headers.has('User-Agent')) {
-		headers.set('User-Agent', 'node-fetch/1.0 (+https://github.com/bitinn/node-fetch)');
+		headers.set('User-Agent', 'node-fetch');
 	}
 
 	// HTTP-network-or-cache fetch step 2.15
@@ -252,9 +246,9 @@ export function getNodeRequestOptions(request) {
 		query: parsedURL.query,
 		href: parsedURL.href,
 		method: request.method,
-		headers: exportNodeCompatibleHeaders(headers),
+		headers: headers[Symbol.for('nodejs.util.inspect.custom')](),
 		agent
 	};
 
 	return requestOptions;
-}
+};